Another major rewrite.
Moving to a new art style. Top-front angle view.

Map data is now loaded from a separate file made using the program
Tiled.

This is still a work in progress.
It’s based on the tutorial from http://kieve.ca/?p=10 but requires
major rewrites and restructuring. This is just the start.